rust_i18n_macro

Macro i18n

i18n!() { /* proc-macro */ }

Init I18n translations.

This will load all translations by glob **/*.yml from the given path, default: ${CARGO_MANIFEST_DIR}/locales.

Attributes

fallback for set the fallback locale, if present t! macro will use it as the fallback locale.

backend for set the backend, if present t! macro will use it as the backend.

metadata to enable/disable loading of the [package.metadata.i18n] config from Cargo.toml, default: true.

minify_key for enable/disable minify key, default: DEFAULT_MINIFY_KEY.

minify_key_len for set the minify key length, default: DEFAULT_MINIFY_KEY_LEN,

The range of available values is from 0 to 24.

minify_key_prefix for set the minify key prefix, default: DEFAULT_MINIFY_KEY_PREFIX.

minify_key_thresh for set the minify key threshold, default: DEFAULT_MINIFY_KEY_THRESH.

If the length of the value is less than or equal to this value, the value will not be minified.

i18n!();
i18n!("locales");
i18n!("locales", fallback = "en");
i18n!("locales", fallback = ["en", "es"]);
i18n!("locales", fallback = ["en", "es"],
      minify_key = true,
      minify_key_len = 12,
      minify_key_prefix = "t_",
      minify_key_thresh = 64);
i18n!();